+++ src/Client/ClientVersion.cs
@@ -7,8 +7,8 @@ public static class ClientVersion
     public static Version Value =>
         Assembly.GetExecutingAssembly()?.GetName().Version ?? new Version("0.0.0.0");
 
-    public static bool IsCompatibleWithServerVersion(Version serverVersion) => 
-        Value.ComparableVersionNumber() == serverVersion.ComparableVersionNumber();
+    public static bool IsCompatibleWithServerVersion(this string serverVersion) => 
+        VersionNumber.StartsWith(serverVersion);
 
     public static string VersionNumber => Value.ThreeDigitsVersionNumber();
 }