+++ src/Clock/ClockServerOptions.cs
@@ -19,14 +19,6 @@ public sealed class ClockServerOptions
     public TimeSpan StartTime { get; set; }
     public TimeSpan Duration { get; set; }
     public double Speed { get; set; }
-    public SoundOptions Sounds { get; set; } = new();
     public string TimeZoneId { get; set; } = "Central Europe Standard Time";
 }
 
-
-public sealed class SoundOptions
-{
-    public bool PlayAnnouncements { get; set; }
-    public string? StartSoundFilePath { get; set; }
-    public string? StopSoundFilePath { get; set; }
-}